pyproject.toml
[tool.poetry]
name = "test-game-builds"
version = "0.1.0"
description = "Test task for Saber team"
authors = ["lisvv <81390205+lisvv@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.5"
fastapi = "0.95.2"
uvicorn = "0.22.0"
pyyaml = "6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
